Added PowerUp/Coin Lists to Launch
Also Added some music Files to Rabbit Game

+++ Games/Launch/Assets/Scripts/Power Up Scripts/StaticVariables.cs
@@ -1,5 +1,5 @@
 ﻿public class StaticVariables{
 
-	public static int totalCoins = 0;
+	public static float totalCoins = 0f;
 
 }
